requirements.txt
speechrecognition
pyttsx3
openai
Pillow